+++ R/fit_c4_aci.R
@@ -81,7 +81,7 @@ fit_c4_aci <- function(
     required_variables <- list()
     required_variables[[ca_column_name]] <- unit_dictionary('Ca')
 
-    check_required_variables(replicate_exdf, required_variables)
+    check_required_variables(replicate_exdf, required_variables, check_NA = FALSE)
 
     # Assemble lower, upper, and fit_options
     luf <- assemble_luf(
@@ -405,10 +405,11 @@ fit_c4_aci <- function(
     replicate_identifiers[, 'c4_assimilation_msg'] <- replicate_exdf[1, 'c4_assimilation_msg']
 
     # Store the results
-    replicate_identifiers[, 'operating_Ci']       <- operating_point_info$operating_Ci
-    replicate_identifiers[, 'operating_PCm']      <- operating_point_info$operating_PCm
-    replicate_identifiers[, 'operating_An']       <- operating_point_info$operating_An
-    replicate_identifiers[, 'operating_An_model'] <- operating_An_model
+    replicate_identifiers[, 'operating_Ci']        <- operating_point_info$operating_Ci
+    replicate_identifiers[, 'operating_PCm']       <- operating_point_info$operating_PCm
+    replicate_identifiers[, 'operating_An']        <- operating_point_info$operating_An
+    replicate_identifiers[, 'operating_An_model']  <- operating_An_model
+    replicate_identifiers[, 'operating_point_msg'] <- operating_point_info$operating_point_msg
 
     # Get an updated likelihood value using the RMSE
     replicate_identifiers[, 'optimum_val'] <- if (fit_failure) {
@@ -457,6 +458,7 @@ fit_c4_aci <- function(
         c('estimate_operating_point', 'operating_Ci',        replicate_exdf$units[[ci_column_name]]),
         c('estimate_operating_point', 'operating_PCm',       replicate_exdf$units[[pcm_column_name]]),
         c('estimate_operating_point', 'operating_An',        replicate_exdf$units[[a_column_name]]),
+        c('estimate_operating_point', 'operating_point_msg', ''),
         c('fit_c4_aci',               'operating_An_model',  replicate_exdf$units[[a_column_name]]),
         c('fit_c4_aci',               'convergence',         ''),
         c('fit_c4_aci',               'convergence_msg',     ''),