+++ tests/testthat/test-fit_c3_variable_j.R
@@ -79,16 +79,23 @@ test_that('fit results have not changed (no alpha)', {
     # default optimizer
     set.seed(1234)
 
-    fit_res <- fit_c3_variable_j(
-        one_curve,
-        Ca_atmospheric = 420,
-        fit_options = list(alpha_old = 0, alpha_g = 0, alpha_s = 0),
-        optim_fun = optimizer_deoptim(200),
-        require_positive_gmc = 'all',
-        hard_constraints = 2,
-        calculate_confidence_intervals = TRUE,
-        remove_unreliable_param = 2,
-        check_j = FALSE
+    # Here we will also check that NA values of Ca do not cause fit failures
+    one_curve_weird <- one_curve
+
+    one_curve_weird[, 'Ca'] <- NA
+
+    fit_res <- expect_silent(
+        fit_c3_variable_j(
+            one_curve_weird,
+            Ca_atmospheric = 420,
+            fit_options = list(alpha_old = 0, alpha_g = 0, alpha_s = 0),
+            optim_fun = optimizer_deoptim(200),
+            require_positive_gmc = 'all',
+            hard_constraints = 2,
+            calculate_confidence_intervals = TRUE,
+            remove_unreliable_param = 2,
+            check_j = FALSE
+        )
     )
 
     fit_res$parameters <- calculate_temperature_response(
@@ -127,8 +134,13 @@ test_that('fit results have not changed (no alpha)', {
 
     expect_equal(
         as.numeric(fit_res$parameters[1, c('operating_Ci', 'operating_Cc', 'operating_An', 'operating_An_model')]),
-        c(294.70316, 216.41088, 37.51608, 40.05385),
+        as.numeric(c(NA, NA, NA, NA)),
         tolerance = TOLERANCE
+    )
+
+    expect_equal(
+        fit_res$parameters[1, 'operating_point_msg'],
+        'All values of the atmospheric CO2 concentration column (Ca) are NA'
     )
 
     expect_equal(
@@ -191,6 +203,17 @@ test_that('fit results have not changed (alpha_old)', {
         as.numeric(fit_res$parameters[1, c('Vcmax_at_25_upper', 'J_at_25_upper', 'RL_at_25_upper', 'tau_upper', 'Tp_at_25_upper')]),
         c(250.388, 258.740, 1.912, 0.412, Inf),
         tolerance = TOLERANCE
+    )
+
+    expect_equal(
+        as.numeric(fit_res$parameters[1, c('operating_Ci', 'operating_Cc', 'operating_An', 'operating_An_model')]),
+        c(294.70316, 213.83837, 37.51608, 40.03141),
+        tolerance = TOLERANCE
+    )
+
+    expect_equal(
+        fit_res$parameters[1, 'operating_point_msg'],
+        ''
     )
 
     expect_equal(